+++ automatic/io-ninja/Changelog.md
@@ -1,5 +1,8 @@
 # ![IO Ninja Changelog](https://img.shields.io/badge/IO%20Ninja-Package%20Changelog-blue.svg?style=for-the-badge)
 
+## UPCOMING
+- **IMPROVEMENT:** Update install/uninstall scripts to latest edition
+
 ## Version: 3.7.4 (2017-06-21)
 - **ENHANCEMENT:** Added changelog for package
 - **ENHANCEMENT:** Extracted Description to its own file